add channel partner map and rename file
