+++ src/sentry/auth/partnership_configs.py
