+++ tests/sentry/auth/providers/fly/test_provider.py
@@ -1,7 +1,7 @@
 import pytest
 
 from sentry.auth.exceptions import IdentityNotValid
-from sentry.auth.partnership_config import ChannelName
+from sentry.auth.partnership_configs import ChannelName
 from sentry.auth.provider import MigratingIdentityId
 from sentry.models import AuthIdentity, AuthProvider
 from sentry.testutils import TestCase
